server/requirements.txt
flask==3.0.3
requests==2.32.3
qrcode==7.4.2
Pillow==10.4.0

